requirements.txt
boto3
pandas
requests
pyyaml
